Move extensions manager to folder
