+++ api/src/extensions/index.ts
@@ -1,0 +1,13 @@
+import { ExtensionManager } from './manager.js';
+
+let extensionManager: ExtensionManager | undefined;
+
+export function getExtensionManager(): ExtensionManager {
+	if (extensionManager) {
+		return extensionManager;
+	}
+
+	extensionManager = new ExtensionManager();
+
+	return extensionManager;
+}
