+++ api/src/extensions/types.ts
@@ -1,0 +1,16 @@
+import type { EndpointConfig, HookConfig, OperationApiConfig } from '@directus/types';
+
+export type BundleConfig = {
+	endpoints: { name: string; config: EndpointConfig }[];
+	hooks: { name: string; config: HookConfig }[];
+	operations: { name: string; config: OperationApiConfig }[];
+};
+
+export type AppExtensions = string | null;
+
+export type ApiExtensions = { path: string }[];
+
+export interface Options {
+	schedule: boolean;
+	watch: boolean;
+}
